NX-2517: drop assumption that CPUs in /proc/stat are consequtive
Previously, the count of "cpuN" lines in /proc/stat was used to allocate
memory "tables" for per-core metrics. This is not safe since the same
code uses the CPU index (the N from the "cpuN" line) to calculate the
offset in the table. This is because the CPU core indices are not
necessarily sequential or start with zero in /proc/stat. This assumption
is incurrect when some CPUs are explicitly disabled, like with

    echo 0 > /sys/devices/system/cpu/cpu1/online

or when running under RR debugger (RR manipulates the tracee into
believing there's only one core, to reduce the number of threads).

Violating this faulty assumption leads CpuUsageCollector() to perform
many out-of-bounds writes, causing memory corruption and potential
crashes.

This commit makes StartCpuUsageCollector() allocate stats tables based
on the highest CPU core index found in /proc/stat .

We still assume that the "cpuN" lines go in ascending order.

This commit does not address the situation when CpuUsageCollector()
sees newly-enabled CPU cores after the tables were allocated.

NX-2517: subagents/linux/cpu.cpp: enforce locking

Revise locking assuming that allocated "tables" of measurements may
change anytime. This is so that in the following commit we introduce
handling of the change of the number of CPU cores visible to the agent.

NX-2517: rework linux/cpu.cpp into semantic C++

The task of handling changes in presence of CPU cores poses challenge in
ensuring the reliability of implementation while not overinvesting.

This commit changes low level C-style code into C++ which spells out the
concepts in use, gives them names to avoid confusion.

E.g. overall CPU measurements are no longer an index zero in a
memory allocation with (cpu + 1) chunks, but its own object.

NX-2517: add tests for CPU usage collector

This test script was used to aid debugging the overhaul of
linux/cpu.cpp. It doesn't fit into any automated suite and may be not
worth merging, but it's added here for historical purposes.

To test actual CPU cores set changing synamically, do the following:
in a virtual machine, while running this test, run:

    while true; do for core in $(seq 1 $(( "$(nproc --all)" - 1 )) ); do echo "echo $(( RANDOM % 2 )) > /sys/devices/system/cpu/cpu$core/online"; done | bash -x; done

NX-2517: add unit test for linux CPU stats collector class