NX-2517: add tests for CPU usage collector
This test script was used to aid debugging the overhaul of
linux/cpu.cpp. It doesn't fit into any automated suite and may be not
worth merging, but it's added here for historical purposes.

To test actual CPU cores set changing synamically, do the following:
in a virtual machine, while running this test, run:

    while true; do for core in $(seq 1 $(( "$(nproc --all)" - 1 )) ); do echo "echo $(( RANDOM % 2 )) > /sys/devices/system/cpu/cpu$core/online"; done | bash -x; done