NX-2455: implement StartTLS for IMAP, POP3
Completeness of protocol response is now checked by regex matching, in
order to reuse the most of connection handling code and keep it simple.

Custom description tag is used to make log statements informative
without a need to pass host, port and protocol string to all the helper
functions.

Added tests for the new protocols.