NX-2488: Restore NXSL callWebService() being non-cached, add caching option
Node::callWebService(), and BaseWebServiceRequestWithData(), already has
two mandatory fixed positional parameters: data and content type.
All the following optional variadic parameters are passed on.

The boolean flag "acceptCached" is added.
It is a named parameter, it can be given on third position. This is
maintains backwards compatibility with existing scripts.